+++ doc/source/changelog.rst
@@ -11,6 +11,7 @@ MontePy Changelog
 **Bugs Fixed**
 
 * Fixed a bug where ``append_renumber`` raised a ``TypeError`` when called with an object whose ``number`` is ``None`` (e.g. an object created with no arguments) (:issue:`880`).
+* ``Cell.universe`` can now be set to ``None`` (or deleted via ``del cell.universe``) to reset the universe assignment back to the default (:issue:`902`).
 
 **Feature Added**